+++ src/index.ts
@@ -1,88 +1,80 @@
-export const REDUX_ATOMIC_ACTION = 'reduxAtomic/REDUX_ATOMIC_ACTION'
-
-export type AtomicReducerFunc<s, t> = (state: s) => s | t
-export type AtomicReducer<s, t> = (state: s, action: AtomicAction<s, t>) => s | t
+export type AtomicReducerFunc<s, t> = (state: s) => s | t;
+export type AtomicReducer<s, t> = (state: s, action: AtomicAction<s, t>) => s | t;
 
 export interface AtomicAction<s, t> {
-    type: string
-    payload: any[]
-    meta: {
-        id: typeof REDUX_ATOMIC_ACTION
-        key: object
-        change: AtomicReducerFunc<s, t>
-    }
+  type: string;
+  payload: any[];
 }
 
+export type f<s, t> = () => AtomicReducerFunc<s, t>;
+export type f1<s, t, A> = (a: A) => AtomicReducerFunc<s, t>;
+export type f2<s, t, A, B> = (a: A, b: B) => AtomicReducerFunc<s, t>;
+export type f3<s, t, A, B, C> = (a: A, b: B, c: C) => AtomicReducerFunc<s, t>;
+export type f4<s, t, A, B, C, D> = (a: A, b: B, c: C, d: D) => AtomicReducerFunc<s, t>;
 
-export type f<s, t> = () => AtomicReducerFunc<s, t>
-export type f1<s, t, A> = (a: A) => AtomicReducerFunc<s, t>
-export type f2<s, t, A, B> = (a: A, b: B) => AtomicReducerFunc<s, t>
-export type f3<s, t, A, B, C> = (a: A, b: B, c: C) => AtomicReducerFunc<s, t>
-export type f4<s, t, A, B, C, D> = (a: A, b: B, c: C, d: D) => AtomicReducerFunc<s, t>
+export type g<s, t> = () => AtomicAction<s, t>;
+export type g1<s, t, A> = (a: A) => AtomicAction<s, t>;
+export type g2<s, t, A, B> = (a: A, b: B) => AtomicAction<s, t>;
+export type g3<s, t, A, B, C> = (a: A, b: B, c: C) => AtomicAction<s, t>;
+export type g4<s, t, A, B, C, D> = (a: A, b: B, c: C, d: D) => AtomicAction<s, t>;
 
-export type g<s, t> = () => AtomicAction<s, t>
-export type g1<s, t, A> = (a: A) => AtomicAction<s, t>
-export type g2<s, t, A, B> = (a: A, b: B) => AtomicAction<s, t>
-export type g3<s, t, A, B, C> = (a: A, b: B, c: C) => AtomicAction<s, t>
-export type g4<s, t, A, B, C, D> = (a: A, b: B, c: C, d: D) => AtomicAction<s, t>
+export type GenericAction<s, t> = (...a: any[]) => AtomicReducerFunc<s, t>;
 
-export function createAtomic<s, t>(initialState: s, name?: string) {
-    const key = {}
-    const reducerName = name || "anon"
+export function createAtomic<s, t>(reducerName: string, initialState: s, reducers: GenericAction<s, t>[]) {
+  const reducerFuncs = saveReducerFuncs(reducers);
+
+  return {
+    reducer,
+    wrap: wrapper
+  };
+
+  function reducer(state: s, action: AtomicAction<s, t>): s | t {
+    const thisState = state || initialState;
+    const params = action && action.payload ? action.payload : [];
+    const funcKey = parseActionKeyFromType(action.type);
+    const func = funcKey in reducerFuncs ? reducerFuncs[funcKey] : false;
+    return !func || !func(...params) ? thisState : func(...params)(thisState);
+  }
+
+  function wrapStateFunc<s, t>(
+    func: AtomicReducerFunc<s, t>,
+    params: any[],
+    actionName: string
+  ): AtomicAction<s, t> {
     return {
-        reducer: createAtomicReducer(initialState, key),
-        key,
-        name: reducerName,
-        wrap: wrapper
-    }
+      type: generateKey(reducerName, actionName),
+      payload: stripUndefined(params)
+    };
+  }
 
-    function createAtomicReducer<s, t>(initialState: s, key: object) {
-        return (state: s, action: AtomicAction<s, t>): s | t => {
-            const thisState = state || initialState
-            return (
-                action.meta &&
-                action.meta.key === key &&
-                action.meta.id === REDUX_ATOMIC_ACTION
-            ) ? action.meta.change(thisState) : thisState
-        }
-    }
+  function stripUndefined(list: any[]): any[] {
+    return list.reduce((acc, val) => {
+      return val !== undefined ? [...acc, val] : [...acc];
+    }, []);
+  }
 
-    function wrapStateFunc<s, t>(func: AtomicReducerFunc<s, t>, params: any[], actionName?: string): AtomicAction<s, t> {
-        return {
-            type: generateKey(reducerName, actionName),
-            payload: stripUndefined(params),
-            meta: {
-                id: REDUX_ATOMIC_ACTION,
-                key,
-                change: func
-            }
-        }
-    }
+  function wrapper<s, t>(func: f<s, t>): g<s, t>;
+  function wrapper<s, t, A>(func: f1<s, t, A>): g1<s, t, A>;
+  function wrapper<s, t, A, B>(func: f2<s, t, A, B>): g2<s, t, A, B>;
+  function wrapper<s, t, A, B, C>(func: f3<s, t, A, B, C>): g3<s, t, A, B, C>;
+  function wrapper<s, t, A, B, C, D>(func: f4<s, t, A, B, C, D>): g4<s, t, A, B, C, D> {
+    return function(a: A, b: B, c: C, d: D) {
+      return wrapStateFunc(func(a, b, c, d), [a, b, c, d], func.name);
+    };
+  }
 
-    function stripUndefined(list: any[]): any[] {
-        return list.reduce((acc, val) => {
-            return (val !== undefined) ? [...acc, val] : [...acc]
-        }, [])
-    }
+  function saveReducerFuncs(reducers: GenericAction<s, t>[]) {
+    return reducers.reduce((acc, reducer) => {
+      const funcName = reducer.name;
+      return { ...acc, [funcName]: reducer };
+    }, {});
+  }
 
-    function wrapper<s, t>(func: f<s, t>, funcName?: string): g<s, t>
-    function wrapper<s, t, A>(func: f1<s, t, A>, funcName?: string): g1<s, t, A>
-    function wrapper<s, t, A, B>(func: f2<s, t, A, B>, funcName?: string): g2<s, t, A, B>
-    function wrapper<s, t, A, B, C>(func: f3<s, t, A, B, C>, funcName?: string): g3<s, t, A, B, C>
-    function wrapper<s, t, A, B, C, D>(func: f4<s, t, A, B, C, D>, funcName?: string): g4<s, t, A, B, C, D> {
-        return function (a: A, b: B, c: C, d: D) {
-            return wrapStateFunc(func(a, b, c, d), [a, b, c, d], funcName)
-        }
-    }
+  function parseActionKeyFromType(actionType: string): string {
+    return actionType.includes(reducerName) ? actionType.substr(actionType.indexOf("_") + 1) : "";
+  }
 
-    function generateKey(identifier: string, actionName?: string): string {
-        return actionName || 'ATOMIC_' + identifier
-    }
+  function generateKey(reducerName: string, actionName: string): string {
+    return reducerName + "_" + actionName;
+  }
 }
-
-
-
-
-
-
-
